Added file bugfix_3.0_preparition
This file contains fixes for delete a book option. and display full library option.

Now the view of books is updated after adding a new book in a list.
In display full library option, the label "your library is empty" after adding a new book is not visible anymore. 

In to do list, there are still some bugfixes and features to repair and/or implement.